Fix rounding bug and increase performance radically with kissfft and pairs of short files
If both input files are similar in length, but precisely one is
odd-length, the convolution might fail due to fftlen being odd. This was
fixed by creating a "safe size" using factors of 2,3,4,5 that is only a
small amount larger than the target fftsize, rather than the power of 2,
which may waste space and time.

This also happens to make some convolutions MUCH faster with kissfft,
which is bad at handling large prime factors in its fft routines. The
test files went from 3 minute convolution to under a half second
convolution. This is not an average case.